+++ src/Recyclarr/Command/IRadarrCommand.cs
@@ -2,4 +2,5 @@ namespace Recyclarr.Command;
 
 public interface IRadarrCommand : IServiceCommand
 {
+    bool ListCustomFormats { get; }
 }