+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Radarr: New `--list-custom-formats` CLI option for getting a flat list of all CFs in the guide in
+  YAML format, ready to copy & paste.
+
 ## [2.2.1] - 2022-06-18
 
 ### Changed